Don't use library-only flag
As far as I can tell, its only effect is to not build the executable.
But, we're passing `-fobject-code` to GHCi anyway, so it's not going to
build the executable. And `app/main.hs` is main for the executable part
of this package, so ghcid (or something) isn't happy about us setting
`app/main.hs` as the target when it's excluded by the flags.

As far as I can tell, it doesn't have any performance impact, so we can
skip it and make ghci happy.